Elevate the chord type
